requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
cors==1.0.1
coverage==7.2.7
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
gevent==23.7.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
midjourney-api==1.0.7
multidict==6.0.4
openai==0.27.8
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PySocks==1.7.1
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-decouple==3.5
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
